+++ jAssert.cpp
@@ -12,11 +12,11 @@
 static BaseAssertionHandler defaultHandler;
 BaseAssertionHandler *assertionHandler = &defaultHandler;
 
-void BaseAssertionHandler::AssertionFailed(long line, const char *function, const char *file)
+void BaseAssertionHandler::AssertionFailed(int line, const char *function, const char *file)
 {
 	// Report the error
 	// This code was moved out of assertion macro for code brevity and to make modification easier
-	DebugPrintfFatal("An assertion was failed and the program has crashed", "Assertion failed on line %ld, function %s, file %s\n", line, function, file);
+	DebugPrintfFatal("An assertion was failed and the program has crashed", "Assertion failed on line %d, function %s, file %s\n", line, function, file);
 	PullDownCode();
 	// Included to satisfy the compiler, which wants to see unambiguously that this function will never return
 	assert(false);
@@ -33,18 +33,18 @@ void BaseAssertionHandler::PullDownCode(void)
 	// Now trigger the crash by dereferencing a null pointer
 	// Note that the analyzer doesnt like this, so we hide it from the analyzer
 #ifndef __clang_analyzer__
-	*((volatile long *)0L) = 0;
+	*((volatile int *)0L) = 0;
 #endif
 	// Included to satisfy the compiler, which wants to see unambiguously that this function will never return
 	assert(false);
 }
 
-bool BaseAssertionHandler::CheckCondition(bool condition, long line, const char *function, const char *file)
+bool BaseAssertionHandler::CheckCondition(bool condition, int line, const char *function, const char *file)
 {
 	if (!condition)
 	{
 		// Report the error
-		DebugPrintf("Check failed on line %ld, function %s, file %s\n", line, function, file);
+		DebugPrintf("Check failed on line %d, function %s, file %s\n", line, function, file);
 	}
 	return condition;
 }
